# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hommies Group Api Documentation",
        "contact": {
            "name": "developer",
            "email": "rohit7305.rk@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        },
        {
            "url": "http://localhost/hommiesGroup/api",
            "description": "local server"
        },
        {
            "url": "https://hommiesgroup.com/api",
            "description": "Live server"
        }
    ],
    "paths": {
...
        "/user/register": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "user register",
                "description": "user register",
                "operationId": "userRegister",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Validator Error"
                    },
                    "401": {
                        "description": "Authentication Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Something went wrong"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User register successfully"
                                        },
                                        "user": {
                                            "type": "string",
                                            "example": "User details"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "basicAuth": []
                    }
                ]
            }
        },
...
        "/user/login": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "user login",
                "description": "user login",
                "operationId": "userLogin",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "device_token",
                                    "device_name",
                                    "device_type"
                                ],
                                "properties": {
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password2"
                                    },
                                    "email": {
                                        "type": "email",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "device_token": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "DVtoken"
                                    },
                                    "device_name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "DVname"
                                    },
                                    "device_type": {
                                        "type": "integer",
                                        "format": "string",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Validator Error"
                    },
                    "401": {
                        "description": "Authentication Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Something went wrong"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "login successfully"
                                        },
                                        "user": {
                                            "type": "string",
                                            "example": "User details"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "basicAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}